% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIDmappingEfficiency.R
\name{checkIDmappingEfficiency}
\alias{checkIDmappingEfficiency}
\title{Checks mapping efficiency for IDs on the most commonly used ID types}
\usage{
checkIDmappingEfficiency(IDs, keytype, db = org.Hs.eg.db)
}
\arguments{
\item{IDs}{character vector of IDs}

\item{keytype}{corresponding keytype}

\item{db}{database object, default org.Hs.eg.db}
}
\value{
Returns a data.frame containing information of mapping efficiency on GO, ENTREZ and KEGG IDs
}
\description{
Checks mapping efficiency for IDs on the most commonly used ID types
}
\examples{
data("exampleContrastData", package = "fgczgseaora")
df <- get_UniprotID_from_fasta_header(exampleContrastData)
checkIDmappingEfficiency(df$UniprotID, keytype = "UNIPROT")

}
