% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteHelpers.R
\name{copy_bfabric_2grpGSEA}
\alias{copy_bfabric_2grpGSEA}
\title{copy all files needed to generate a bfabric report}
\usage{
copy_bfabric_2grpGSEA(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy all files needed to generate a bfabric report
}
