% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteHelpers.R
\name{copy_bfabric_ClusterProfiler}
\alias{copy_bfabric_ClusterProfiler}
\title{copy all files needed to generate a bfabric report for cluster profiling}
\usage{
copy_bfabric_ClusterProfiler(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy all files needed to generate a bfabric report for cluster profiling
}
