% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteHelpers.R
\name{copy_gsea_report}
\alias{copy_gsea_report}
\title{copy all files needed to generate GSEA report.}
\usage{
copy_gsea_report(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy all files needed to generate GSEA report.
}
