% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteHelpers.R
\name{copy_multigroup_GSEA}
\alias{copy_multigroup_GSEA}
\title{copy multigroup analysis}
\usage{
copy_multigroup_GSEA(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy multigroup analysis
}
