% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignetteHelpers.R
\name{copy_ora_report}
\alias{copy_ora_report}
\title{copy all files needed to generate ORA report}
\usage{
copy_ora_report(workdir = getwd())
}
\arguments{
\item{workdir}{directory where to copy file - default is current working directory.}
}
\description{
copy all files needed to generate ORA report
}
