% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_addurls}
\alias{cp_addurls}
\title{add uris to to cp table outputs}
\usage{
cp_addurls(
  data,
  caption,
  coreEnrichment = "core_enrichment",
  signif2 = c("NES", "pvalue", "FDR"),
  gocarts = "http://amigo.geneontology.org/amigo/term/",
  genecarts = "https://www.ncbi.nlm.nih.gov/gene/"
)
}
\description{
add uris to to cp table outputs
}
