% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_addurls_GO}
\alias{cp_addurls_GO}
\title{add uris to to cp table outputs}
\usage{
cp_addurls_GO(
  data,
  caption,
  coreEnrichment = "core_enrichment",
  signif2 = c("NES", "pvalue", "FDR"),
  gocarts = "http://amigo.geneontology.org/amigo/term/",
  genecarts = "https://www.ncbi.nlm.nih.gov/gene/"
)
}
\arguments{
\item{data}{dataframe}

\item{caption}{table caption}

\item{coreEnrichment}{column with core enrichments}

\item{signif2}{which columns contains digits to round to significance 2.}

\item{gocarts}{default http://amigo.geneontology.org/amigo/term/}

\item{genecarts}{default https://www.ncbi.nlm.nih.gov/gene/}
}
\description{
add uris to to cp table outputs
}
\seealso{
Other clusterProfiler: 
\code{\link{cp_clusterDPAEuclideanDist}()},
\code{\link{cp_clusterHClustEuclideanDistDeepslit}()},
\code{\link{cp_clusterHClustEuclideanDist}()},
\code{\link{cp_compClust}()},
\code{\link{cp_filterforNA}()}
}
\concept{clusterProfiler}
