% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_clusterHClustEuclideanDist}
\alias{cp_clusterHClustEuclideanDist}
\title{cluster using hclust}
\usage{
cp_clusterHClustEuclideanDist(x, nrCluster, method = "complete")
}
\description{
cluster using hclust
}
