% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_compClust}
\alias{cp_compClust}
\title{same as clusterProfiler::compareCluster but with defaults and error handling}
\usage{
cp_compClust(geneClusters, orgDB, universe, ont = "BP")
}
\arguments{
\item{geneClusters}{a list of entrez gene id.}

\item{orgDB}{organizm db}

\item{universe}{see \code{\link[clusterProfiler]{compareCluster}}}

\item{ont}{see \code{\link[clusterProfiler]{compareCluster}}}
}
\description{
same as clusterProfiler::compareCluster but with defaults and error handling
}
\seealso{
Other clusterProfiler: 
\code{\link{cp_addurls_GO}()},
\code{\link{cp_clusterDPAEuclideanDist}()},
\code{\link{cp_clusterHClustEuclideanDistDeepslit}()},
\code{\link{cp_clusterHClustEuclideanDist}()},
\code{\link{cp_filterforNA}()}
}
\concept{clusterProfiler}
