% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_filterforNA}
\alias{cp_filterforNA}
\title{remove proteins with more NA's than in 60\% of samples}
\usage{
cp_filterforNA(mdata)
}
\arguments{
\item{mdata}{data matrix}
}
\description{
remove proteins with more NA's than in 60\% of samples
}
\seealso{
Other clusterProfiler: 
\code{\link{cp_addurls_GO}()},
\code{\link{cp_clusterDPAEuclideanDist}()},
\code{\link{cp_clusterHClustEuclideanDistDeepslit}()},
\code{\link{cp_clusterHClustEuclideanDist}()},
\code{\link{cp_compClust}()}
}
\concept{clusterProfiler}
