% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dist_JK}
\alias{dist_JK}
\title{JK resampling for dist function}
\usage{
dist_JK(m, method = "euclidean", p = 2)
}
\arguments{
\item{m}{matrix with data}

\item{method}{distance method default euclidean}

\item{p}{power of minkowski distance}
}
\description{
JK resampling for dist function
}
