% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{.runGSEA}
\alias{.runGSEA}
\title{workflow for GSEA}
\usage{
.runGSEA(data, fpath, ID_col = "UniprotID", fc_col = "estimate",
  organism = "hsapiens", target = "geneontology_Biological_Process",
  nperm = 10, outdir = "GSEA")
}
\arguments{
\item{data}{data}

\item{fpath}{file path to write to}

\item{ID_col}{column name containing IDs}

\item{fc_col}{column name containing estimates}

\item{organism}{organism}

\item{target}{target database, default: geneontology_Biological_Process}

\item{nperm}{number of permutations}

\item{outdir}{output directory}
}
\description{
workflow for GSEA
}
