% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{.runSIGORA}
\alias{.runSIGORA}
\title{workflow for sigORA}
\usage{
.runSIGORA(data, target = "GO", fc_col = "estimate",
  ID_col = "UniprotID", fc_threshold = 0.5, outdir = "sigORA",
  greater = TRUE)
}
\arguments{
\item{data}{data}

\item{target}{target database, default: "GO"}

\item{fc_col}{column name containing estimates}

\item{ID_col}{column name containing IDs}

\item{fc_threshold}{threshold for estimate}

\item{outdir}{output directory}

\item{greater}{flag whether to filter > threshold or < threshold}
}
\description{
workflow for sigORA
}
