% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DThelper.R
\name{DT_makeURLfor}
\alias{DT_makeURLfor}
\title{create url for data table}
\usage{
DT_makeURLfor(x, path)
}
\arguments{
\item{x}{eg. uniprot id's P94485}

\item{path}{e.g. https://www.uniprot.org/uniprot/}
}
\description{
create url for data table
}
