% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleSymbols}
\alias{exampleSymbols}
\title{Examplary gene symbols}
\format{A character vector containing 2292 gene symbols}
\usage{
data(exampleSymbols)
}
\description{
Gene symbols for running examples to produce GPS repositories using \code{\link{makeGPS_wrappR}}
}
