% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleUniprotIDs}
\alias{exampleUniprotIDs}
\title{Examplary Uniprot IDs}
\format{A character vector containing 2292 Uniprot IDs}
\usage{
data(exampleUniprotIDs)
}
\description{
Gene symbols for running examples to produce GPS repositories using \code{\link{makeGPS_wrappR}}
}
