% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{fgsea_leading_edge_too_char}
\alias{fgsea_leading_edge_too_char}
\title{convert leading edge column to char for writing to file}
\usage{
fgsea_leading_edge_too_char(xdata, column = "leadingEdge")
}
\description{
convert leading edge column to char for writing to file
}
