% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{fgsea_msigdb}
\alias{fgsea_msigdb}
\title{converts msigdb geneset to fgsea compatible}
\usage{
fgsea_msigdb(msigdbgeneset)
}
\arguments{
\item{msigdbgeneset}{msigdb geneset}
}
\description{
converts msigdb geneset to fgsea compatible
}
