% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{fgsea_msigdb_collections}
\alias{fgsea_msigdb_collections}
\title{retrieve several msigdbr collections as fgsea compatible lists}
\usage{
fgsea_msigdb_collections(msigCollection, species = "Homo sapiens")
}
\description{
retrieve several msigdbr collections as fgsea compatible lists
}
