% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{fgsea_rank_contrasts}
\alias{fgsea_rank_contrasts}
\title{convert data frame into list of ranks}
\usage{
fgsea_rank_contrasts(df, ids = "X1", score = "X2", contrast = "contrast")
}
\description{
convert data frame into list of ranks
}
