% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniprotFromFastaHeader.R
\name{get_UniprotID_from_fasta_header}
\alias{get_UniprotID_from_fasta_header}
\title{Translate Fasta header to UniprotSwissprot ID}
\usage{
get_UniprotID_from_fasta_header(df, idcolumn = "protein_Id")
}
\arguments{
\item{df}{\code{data.frame} with FASTA headers in first column}

\item{idcolumn}{Column name containing the IDs}
}
\value{
Returns the whole \code{data.frame} with a column called \code{UniprotID} containing UniprotSwissprot IDs instead of Fasta headers. This format is easier to use for subsequent ID mappings using the \code{Annotationdbi} package.
}
\description{
Translate Fasta header to UniprotSwissprot ID
}
\examples{
data("exampleContrastData", package = "fgczgseaora")
get_UniprotID_from_fasta_header(exampleContrastData)
}
