% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigDB_helper.R
\name{getMsigdbGenesets}
\alias{getMsigdbGenesets}
\title{getMsigdbGenesets}
\usage{
getMsigdbGenesets(msigCollection, species)
}
\description{
getMsigdbGenesets
}
\examples{
#hallmark$gs_subcat <- "HALLMARK"
#library(msigdbr)

#C5 <- dplyr::bind_rows( msigdbr::msigdbr_collections() \%>\% filter(gs_cat == "C5") \%>\% filter(grepl("^GO:", gs_subcat)),
#                 hallmark,
#                 msigdbr_collections() \%>\% filter(gs_subcat == "CP:KEGG") )
#species <- "Mus musculus"
}
