% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbolFromFasta.R
\name{getSymbolFromFasta}
\alias{getSymbolFromFasta}
\title{Translate Fasta header to gene Symbol}
\usage{
getSymbolFromFasta(df)
}
\arguments{
\item{df}{\code{data.frame} with column called \code{protein_Id}}
}
\value{
Returns the whole \code{data.frame} with a column called \code{Symbol} containing gene symbols instead of Fasta headers. This format is easier to use for subsequent ID mappings using the \code{Annotationdbi} package.
}
\description{
Translate Fasta header to gene Symbol
}
\examples{
data("exampleContrastData", package = "fgczgseaora")
getSymbolFromFasta(exampleContrastData)
}
