% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSymbolFromFasta.R
\name{getSymbolFromSwissprotID}
\alias{getSymbolFromSwissprotID}
\title{Translate Fasta header to gene Symbol}
\usage{
getSymbolFromSwissprotID(df)
}
\arguments{
\item{df}{\code{data.frame} with FASTA headers in first column}
}
\value{
Returns the whole \code{data.frame} with a column called \code{Symbol} containing gene symbols instead of Fasta headers. This format is easier to use for subsequent ID mappings using the \code{Annotationdbi} package.
}
\description{
Translate Fasta header to gene Symbol
}
\examples{
data("exampleContrastData", package = "fgczgseaora")
getSymbolFromSwissprotID(exampleContrastData)
}
