% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniprotFromFastaHeader.R
\name{getUniprotFromFastaHeader}
\alias{getUniprotFromFastaHeader}
\title{Translate Fasta header to UniprotSwissprot ID}
\usage{
getUniprotFromFastaHeader(df)
}
\arguments{
\item{df}{\code{data.frame} with FASTA headers in first column}
}
\value{
Returns the whole \code{data.frame} with a column called \code{UniprotID} containing UniprotSwissprot IDs instead of Fasta headers. This format is easier to use for subsequent ID mappings using the \code{Annotationdbi} package.
}
\description{
Translate Fasta header to UniprotSwissprot ID
}
\examples{
data("exampleContrastData", package = "fgczgseaora")
getUniprotFromFastaHeader(exampleContrastData)
}
