% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSEA}
\alias{GSEA}
\title{Exemplary GSEA output}
\format{A list containing the 5 entries as described below.}
\usage{
data(GSEA)
}
\description{
GSEA output from the \code{run_GSEA.R} script provided in the \code{inst/RunScripts/} directory
}
\details{
\itemize{
  \item \code{organism.} The organism used in the analysis
  \item \code{target.} Target database
  \item \code{outputDir.} Output directory of where the results are stores
  \item \code{input_data.} Input data containing gene symbols and effect estimates
  \item \code{GSEA_res.} Results from \code{{WebGestaltR::WebGestaltR}}
}
}
