% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGPS_wrappR.R
\name{makeGPS_wrappR}
\alias{makeGPS_wrappR}
\title{makeGPS Wrapper}
\usage{
makeGPS_wrappR(ids, target = "KEGG")
}
\arguments{
\item{ids}{Character vector of gene symbols (experimental background)}

\item{target}{Character, database to be used to generate GPS repository ("KEGG", "GO", "reactome")}
}
\value{
A GPS repository based on the self specified background in \code{ids}, see \code{sigora::makeGPS()} for more information.
}
\description{
Function for generating background GPS repository for sigora and ora
}
\examples{
data("exampleSymbols", package = "fgczgseaora")
myGPSrepo <- makeGPS_wrappR(ids = exampleSymbols, target = "KEGG")

}
\seealso{
\code{sigora::makeGPS()}
}
