% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ids_uniprot.R
\name{map_ids_2ways}
\alias{map_ids_2ways}
\title{map ids two ways}
\usage{
map_ids_2ways(
  clusterAssignment,
  ID_col = "UniprotID",
  species = c("Homo sapiens", "Mus musculus")
)
}
\description{
map ids two ways
}
