% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ids_uniprot.R
\name{map_ids_annotationHub}
\alias{map_ids_annotationHub}
\title{map id Annotation Hub}
\usage{
map_ids_annotationHub(
  x,
  ID_col = "UniprotID",
  species = c("Homo sapiens", "Mus musculus")
)
}
\description{
map id Annotation Hub
}
\examples{

library(tidyverse)
fc_estimates <- prora::exampleContrastData
#fc_estimates \%>\% filter(!(grepl("^REV_", protein_Id) | grepl("^zz", protein_Id) | grepl("^CON__", protein_Id)))
fc_estimates <- fc_estimates \%>\% filter(!(grepl("^REV_|^zz|^CON__", protein_Id)))
filtered_dd <- get_UniprotID_from_fasta_header(fc_estimates, idcolumn = "protein_Id")
dim(filtered_dd)
tmp <- map_ids_annotationHub(filtered_dd, species = "Homo sapiens")
sum(is.na(tmp$P_ENTREZGENEID ))

}
