% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ids_uniprot.R
\name{map_ids_uniprot}
\alias{map_ids_uniprot}
\title{map_ids_uniprot}
\usage{
map_ids_uniprot(data, ID_col = "UniprotID", from = "ACC+ID",
  to = "P_ENTREZGENEID", format = "tab")
}
\arguments{
\item{data}{data}

\item{ID_col}{column name containing id's}

\item{from}{id type}

\item{to}{id type}

\item{format}{in which format to return the mapping}
}
\description{
Access to uniprot web API. For more detail see https://www.uniprot.org/help/uploadlists
}
\examples{
library(tidyverse)
fc_estimates <- fgczgseaora::exampleContrastData

filtered_dd <- getUniprotFromFastaHeader(fc_estimates, idcolumn = "protein_Id")

map_ids_uniprot( filtered_dd )


}
