% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{mapthe}
\alias{mapthe}
\title{helper function mapping clusters}
\usage{
mapthe(scaledM, cluster)
}
\arguments{
\item{matrix}{with rownames which will be mapped}
}
\description{
helper function mapping clusters
}
\seealso{
Other id_mapping: 
\code{\link{map_ids_2ways}()},
\code{\link{map_ids_annotationHub}()},
\code{\link{map_ids_uniprot}()}
}
\concept{id_mapping}
