% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{run_fgsea_for_allContrasts}
\alias{run_fgsea_for_allContrasts}
\title{run for all contrasts}
\usage{
run_fgsea_for_allContrasts(
  allrnk,
  geneSet,
  nperm = 10000,
  minSize = 25,
  maxSize = 500
)
}
\arguments{
\item{allrnk}{- list of rank arrays}
}
\description{
run for all contrasts
}
