% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{run_fgsea_for_allGeneSets}
\alias{run_fgsea_for_allGeneSets}
\title{used to analyse a single contrast with various genesets}
\usage{
run_fgsea_for_allGeneSets(
  allrnk,
  geneSets,
  nperm = 10000,
  minSize = 25,
  maxSize = 500
)
}
\arguments{
\item{allrnk}{ranklist}

\item{geneSets}{list of gene sets}

\item{nperm}{number of permutations}

\item{minSize}{minimum geneset size}

\item{maxSize}{maximum geneset size}
}
\description{
used to analyse a single contrast with various genesets
}
\seealso{
Other fgsea: 
\code{\link{fgsea_leading_edge_too_char}()},
\code{\link{fgsea_msigdb_collections}()},
\code{\link{fgsea_msigdb}()},
\code{\link{fgsea_rank_contrasts}()},
\code{\link{fgsea_rank}()},
\code{\link{run_fgsea_for_allContrasts}()}
}
\concept{fgsea}
