% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSigora.R
\name{runSIGORA}
\alias{runSIGORA}
\title{workflow for sigORA}
\usage{
runSIGORA(data, target = "GO", score_col = "estimate",
  ID_col = "UniprotID", threshold = 0.5, outdir = "sigORA",
  greater = TRUE)
}
\arguments{
\item{data}{data}

\item{target}{target database, default: "GO"}

\item{score_col}{column name containing estimates}

\item{ID_col}{column name containing IDs}

\item{outdir}{output directory}

\item{greater}{flag whether to filter > threshold or < threshold}

\item{fc_threshold}{threshold for estimate}
}
\description{
workflow for sigORA
}
\examples{
fc_estimates <- fgczgseaora::exampleContrastData
filtered_dd <- get_UniprotID_from_fasta_header(fc_estimates, idcolumn = "protein_Id")
runSIGORA(filtered_dd)
}
