% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runWebGestalt.R
\name{runWebGestaltGSEA}
\alias{runWebGestaltGSEA}
\title{workflow for GSEA}
\usage{
runWebGestaltGSEA(data, fpath, ID_col = "UniprotID",
  score_col = "estimate", organism = "hsapiens",
  target = "geneontology_Biological_Process", nperm = 10,
  outdir = "GSEA", interestGeneType = "uniprotswissprot",
  contrast_name = fpath)
}
\arguments{
\item{data}{data}

\item{fpath}{file path to write to}

\item{ID_col}{column name containing IDs}

\item{score_col}{column name containing estimates}

\item{organism}{organism}

\item{target}{target database, default: geneontology_Biological_Process}

\item{nperm}{number of permutations}

\item{outdir}{output directory}

\item{interestGeneType}{what type of identifier default : "uniprotswissprot"}

\item{contrast_name}{for pretty printing}
}
\description{
workflow for GSEA
}
