% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runfunctions.R
\name{runWebGestaltORA}
\alias{runWebGestaltORA}
\title{workflow for ORA}
\usage{
runWebGestaltORA(data, fpath, organism = "hsapiens",
  ID_col = "UniprotID", target = "geneontology_Biological_Process",
  threshold = 0.5, greater = TRUE, nperm = 10, fc_col = "estimate",
  outdir = "WebGestalt_ORA", interestGeneType = "uniprotswissprot")
}
\arguments{
\item{data}{data}

\item{fpath}{file path to write to}

\item{organism}{organism}

\item{ID_col}{column name containing IDs}

\item{target}{target database, default: "geneontology_Biological_Process"}

\item{threshold}{threshold for estimate}

\item{greater}{indicating direction of threshold}

\item{nperm}{number of permutations}

\item{fc_col}{column name containing estimates}

\item{outdir}{output directory}
}
\description{
workflow for ORA
}
