% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sigora_example}
\alias{sigora_example}
\title{Exemplary sigora output}
\format{A list containing 7 entries as described below}
\usage{
data(sigora_example)
}
\description{
sigora output from the \code{\link{sigoraWrappR}} function containing the following entries
}
\details{
\itemize{
  \item \code{sigora.} Result table from \code{{sigora::sigora}}
  \item \code{ora.} Result table from \code{{sigora::ora}}
  \item \code{fc_threshold.} fold change threshold used for the analysis
  \item \code{GPSrepository.} GPS repository used for the analysis, generated by \code{\link{makeGPS_wrappR}}
  \item \code{database.} Target database
  \item \code{data.} Original input data
  \item \code{proteinsAfterFiltering.} The number of proteins meeting the fold change threshold
}
}
