% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigora_heatmap.R
\name{sigora_heatmap}
\alias{sigora_heatmap}
\title{Visualisation of expression patterns in significantly enriched pathways}
\usage{
sigora_heatmap(sigora_res, GPStable)
}
\arguments{
\item{sigora_res}{Object returned by the \code{\link{sigoraWrappR}} function}

\item{GPStable}{Object returned by \code{\link{makeGPS_wrappR}} function, setting \code{dev=TRUE}}
}
\value{
Returns a \code{ggplot} object heatmap
}
\description{
Visualisation of expression patterns in significantly enriched pathways
}
\examples{
data("sigora_example", package = "fgczgseaora")
data("GPStab", package = "fgczgseaora")
sigora_heatmap(sigora_example, GPStab)
}
