% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigoraWrappR.R
\name{sigoraWrappR}
\alias{sigoraWrappR}
\title{sigora Wrapper function}
\usage{
sigoraWrappR(input.file = "", fc_threshold = 0.5, fc_col = "",
  GPSrepos = kegH, df, db = "")
}
\arguments{
\item{input.file}{file path, will be reprinted in the report.}

\item{fc_threshold}{fold change threshold above which (in absolute terms) a protein is considered differentially regulated}

\item{fc_col}{Name of the fold change column, in case the input file contains multiple contrasts}

\item{GPSrepos}{GPS repository used as background, can be generated via \code{sigoraWrappR()}}

\item{df}{data.frame input (at least two columns, first column being gene symbols, other columns being fold changes)}

\item{db}{database used for the generation of the GPS repository}
}
\value{
Returns a \code{list} containing the following elements:
* 1
* 2
* 3
}
\description{
Provides a wrapper for \code{sigora::sigora()} combined with \code{sigora::ora()} and returns a \code{list} used for the generation of an \code{.Rmd} report.
}
\examples{
data("exampleContrastData", package = "fgczgseaora")
df <- getSymbolFromFasta(exampleContrastData)
sigoraWrappR(fc_col = colnames(df)[7], GPSrepos = sigora::kegH, df = df)

}
