% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{webGestaltExample}
\alias{webGestaltExample}
\title{Exemplary WebGestaltR output}
\format{A list of 12 entries}
\usage{
data(webGestaltExample)
}
\description{
Examplary clustering plus ora output using the \code{WebGestaltR} package.
}
