% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebGestaltWrappR.R
\name{webGestaltWrapper}
\alias{webGestaltWrapper}
\title{Wrapper for WebGestaltR overrepresentation analysis in SRMService}
\usage{
webGestaltWrapper(grp2, enrichDatabase, organism, se_threshold,
  work_dir = "output", nrNas, method)
}
\arguments{
\item{grp2}{grp2 S4 object}

\item{enrichDatabase}{Database used for over representation analysis}

\item{organism}{organism proteins in the grp2 object derive from}

\item{se_threshold}{standard error of protein expressions over all conditions. Will be used as a filtering threshold.}

\item{nrNas}{number of allowed NAs per protein. Protein matrix will be filtered using this threshold.}

\item{WebGestaltWrappR}{WebGestaltWrappR}
}
\value{
list with results
}
\description{
Wrapper for WebGestaltR overrepresentation analysis in SRMService
}
