% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_clusterDPAEuclideanDist}
\alias{cp_clusterDPAEuclideanDist}
\title{Cluster using DPA}
\usage{
cp_clusterDPAEuclideanDist(mdata, Z = 1, JK = TRUE)
}
\arguments{
\item{mdata, }{input data}

\item{Z, }{the number of standard deviations fixing the level of
statistical confidence at which one decides to consider
a cluster meaningful. Default value is set to 1.}

\item{JK}{jack knive resampling (default TRUE)}
}
\value{
\itemize{
\item dendrogram - instance of class dendrogram
\item cluster assignments
\item nrCluster number of clusters
}
}
\description{
Cluster using DPA
}
\seealso{
Other clusterProfiler: 
\code{\link{cp_addurls_GO}()},
\code{\link{cp_clusterHClustEuclideanDistDeepslit}()},
\code{\link{cp_clusterHClustEuclideanDist}()},
\code{\link{cp_compClust}()},
\code{\link{cp_filterforNA}()}
}
\concept{clusterProfiler}
