% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_clusterHClustEuclideanDistDeepslit}
\alias{cp_clusterHClustEuclideanDistDeepslit}
\title{cluster using hclust and deepsplit}
\usage{
cp_clusterHClustEuclideanDistDeepslit(x, method = "complete", JK = TRUE)
}
\arguments{
\item{x}{matrix with data}

\item{method}{linkage method (hclust argument)}

\item{JK}{should jack knife resampling be used (defaul TRUE)}
}
\value{
list
\itemize{
\item dendrogram - instance of class dendrogram
\item cluster assignments
\item nrCluster number of clusters
}
}
\description{
cluster using hclust and deepsplit
}
