% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cphelper.R
\name{cp_clusterHClustEuclideanDistDeepslit}
\alias{cp_clusterHClustEuclideanDistDeepslit}
\title{cluster using hclust and deepsplit}
\usage{
cp_clusterHClustEuclideanDistDeepslit(x, method = "complete", JK = TRUE)
}
\description{
cluster using hclust and deepsplit
}
