% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleContrastData}
\alias{exampleContrastData}
\title{Contrast effect estimates}
\format{A \code{data.frame} with 2304 rows and 18 variables}
\usage{
data(exampleContrastData)
}
\description{
Example data set containing effect estimates and p values from different contrasts.
}
