% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{fgsea_msigdb}
\alias{fgsea_msigdb}
\title{converts msigdb geneset to fgsea compatible}
\usage{
fgsea_msigdb(msigdbgeneset)
}
\arguments{
\item{msigdbgeneset}{msigdb geneset}
}
\description{
converts msigdb geneset to fgsea compatible
}
\seealso{
Other fgsea: 
\code{\link{fgsea_leading_edge_too_char}()},
\code{\link{fgsea_msigdb_collections}()},
\code{\link{fgsea_rank_contrasts}()},
\code{\link{fgsea_rank}()},
\code{\link{getMsigdbGenesets}()},
\code{\link{run_fgsea_for_allContrasts}()},
\code{\link{run_fgsea_for_allGeneSets}()}
}
\concept{fgsea}
