% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigDB_helper.R
\name{getMsigdbGenesets}
\alias{getMsigdbGenesets}
\title{getMsigdbGenesets}
\usage{
getMsigdbGenesets(msigCollection, species)
}
\description{
getMsigdbGenesets
}
\examples{

msigdbr::msigdbr_species()
species <- "Homo sapiens"
species <- "Mus musculus"

hallmark <- {msigdbr_collections() \%>\% filter(.data$gs_cat == "H")}

#hallmark$gs_subcat <- "HALLMARK"
C5 <- bind_rows( {msigdbr_collections() \%>\% filter(.data$gs_cat == "C5") \%>\% filter(grepl("^GO:", .data$gs_subcat))},
                 hallmark,
                 {msigdbr_collections() \%>\% filter(.data$gs_subcat == "CP:KEGG")} )

C5
fgseaGSlist <- prora::getMsigdbGenesets(C5, species)

}
