% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPStab}
\alias{GPStab}
\title{Exemplary GPS table}
\usage{
data(GPStab)
}
\description{
GPS table linking GOID, pathway name and genes
\itemize{
  \item \code{pathwayId.} Pathway identifier, in this case a gene ontology id
  \item \code{pathwayName.} Name of the pathway
  \item \code{gene.} Gene symbol
}
}
