% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGPS_wrappR.R
\name{makeGPS_wrappR}
\alias{makeGPS_wrappR}
\title{makeGPS Wrapper
Function for generating background GPS repository for sigora and ora}
\usage{
makeGPS_wrappR(
  ids,
  keytype = "UNIPROT",
  target = c("KEGG", "GO", "reactome"),
  dev = FALSE
)
}
\arguments{
\item{ids}{Character vector of gene symbols (experimental background)}

\item{target}{Character, database to be used to generate GPS repository ("KEGG", "GO", "reactome")}

\item{dev}{Logical, if \code{TRUE} the function returns the \code{data.frame} \code{makeGPS} needs as input}
}
\value{
A GPS repository based on the self specified background in \code{ids}, see \code{\link[sigora]{makeGPS}} for more information.
}
\description{
makeGPS Wrapper
Function for generating background GPS repository for sigora and ora
}
\examples{

data("exampleUniprotIDs", package = "fgczgseaora")
myGPSrepo <- makeGPS_wrappR(ids = exampleUniprotIDs)
myGPSrepo <- makeGPS_wrappR(ids = exampleUniprotIDs, target = "GO")
myGPSrepo <- makeGPS_wrappR(ids = exampleUniprotIDs, target = "react")
}
\seealso{
\code{\link[sigora]{makeGPS}}
}
