% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgsea_utils.R
\name{run_fgsea_for_allContrasts}
\alias{run_fgsea_for_allContrasts}
\title{run for all contrasts}
\usage{
run_fgsea_for_allContrasts(
  allrnk,
  geneSet,
  nperm = 10000,
  minSize = 25,
  maxSize = 500
)
}
\arguments{
\item{allrnk}{list of rank arrays}

\item{geneSet}{single list geneset (e.g. )}

\item{nperm}{number permutations}

\item{minSize}{minimum geneset size}

\item{maxSize}{maximum geneset size}
}
\description{
run for all contrasts
}
\seealso{
Other fgsea: 
\code{\link{fgsea_leading_edge_too_char}()},
\code{\link{fgsea_msigdb_collections}()},
\code{\link{fgsea_msigdb}()},
\code{\link{fgsea_rank_contrasts}()},
\code{\link{fgsea_rank}()},
\code{\link{getMsigdbGenesets}()},
\code{\link{run_fgsea_for_allGeneSets}()}
}
\concept{fgsea}
