% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSigora.R
\name{runSIGORA}
\alias{runSIGORA}
\title{workflow for sigORA}
\usage{
runSIGORA(
  data,
  target = "GO",
  score_col = "estimate",
  ID_col = "UniprotID",
  threshold = 0.5,
  outdir = "sigORA",
  greater = TRUE,
  myGPSrepo = NULL,
  DEBUG = FALSE
)
}
\arguments{
\item{data}{data}

\item{target}{target database, default: "GO"}

\item{score_col}{column name containing estimates}

\item{ID_col}{column name containing IDs}

\item{outdir}{output directory}

\item{greater}{flag whether to filter > threshold or < threshold}

\item{fc_threshold}{threshold for estimate}
}
\description{
workflow for sigORA
}
\examples{

library(fgczgseaora)
fc_estimates <- fgczgseaora::exampleContrastData
filtered_dd <- get_UniprotID_from_fasta_header(fc_estimates, idcolumn = "protein_Id")
myGPSrepo <- makeGPS_wrappR(filtered_dd$UniprotID, target = "GO")

usethis::use_data(myGPSrepo,overwrite = TRUE)
undebug(runSIGORA)
undebug(sigoraWrappR)
res <- runSIGORA(filtered_dd, myGPSrepo = fgczgseaora::myGPSrepo, DEBUG=TRUE)
#rmarkdown::render(res$rmarkdownPath,bookdown::html_document2(number_sections = FALSE),params = res$sigoraData,clean = TRUE)

}
