% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigora_upsetR.R
\name{sigora_upsetR}
\alias{sigora_upsetR}
\title{UpSetR wrapper for sigora results}
\usage{
sigora_upsetR(sigora_res, GPStable, ...)
}
\arguments{
\item{sigora_res}{Object returned by the \code{\link{sigoraWrappR}} function}

\item{GPStable}{Object returned by \code{\link{makeGPS_wrappR}} function, setting \code{dev=TRUE}}

\item{...}{other parameters to \code{upset}}
}
\description{
UpSetR wrapper for sigora results
}
