% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigoraWrappR.R
\name{sigora_upsetR}
\alias{sigora_upsetR}
\title{UpSetR wrapper for sigora results}
\usage{
sigora_upsetR(sigora_res, GPStable, ...)
}
\arguments{
\item{sigora_res}{Object returned by the \code{\link{sigoraWrappR}} function}

\item{GPStable}{Object returned by \code{\link{appR}} function, setting \code{dev=TRUE}}

\item{...}{other parameters to \code{upset}}
}
\description{
UpSetR wrapper for sigora results
}
\examples{
sigora_res <- fgczgseaora::sigora_example
GPStable <- fgczgseaora::GPStab

sigora_upsetR(sigora_res, GPStable)


df <- get_UniprotID_from_fasta_header(fgczgseaora::exampleContrastData)
myGPSrepo <- makeGPS_wrappR(ids = df$UniprotID)
names(myGPSrepo)
res <- sigoraWrappR(df,score_col = "estimate", GPSrepos = myGPSrepo$gps,
                     threshold = 0.5)
sigora_upsetR(res, myGPSrepo$gpsTable)

}
