% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WebGestaltWrappR.R
\name{webGestaltWrapper}
\alias{webGestaltWrapper}
\title{Wrapper for WebGestaltR overrepresentation analysis}
\usage{
webGestaltWrapper(
  quant_data,
  enrichDatabase,
  organism,
  se_threshold,
  work_dir = "output",
  nrNas,
  method,
  nclust = 2
)
}
\arguments{
\item{quant_data}{data.frame of intensity values with protein identifiers as row names}

\item{enrichDatabase}{Database used for over representation analysis}

\item{organism}{organism proteins in the grp2 object derive from}

\item{se_threshold}{standard error of protein expressions over all conditions. Will be used as a filtering threshold.}

\item{work_dir}{output directory to store \code{WebGestaltR} reports}

\item{nrNas}{number of allowed NAs per protein. Protein matrix will be filtered using this threshold.}

\item{method}{clustering method to be used in \code{simpleheatmap3}}

\item{nclust}{number of proteins clusters to consider in the analysis}
}
\value{
list with results
}
\description{
Wrapper for WebGestaltR overrepresentation analysis
}
